requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
streamlit==1.49.0
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.11.7
numpy==2.3.2
faiss-cpu==1.11.0.post1
PyMuPDF==1.26.4
paddleocr==2.9.1
sentence-transformers==5.1.0
langchain-text-splitters==0.3.9
peft==0.18.1
